Add logging for deleting.

+++ Projects/ASI Admin/src/Web Application/Controllers/Show/ExcelUploadController.cs
@@ -227,7 +227,8 @@ namespace asi.asicentral.web.Controllers.Show
                         attendee.EmployeeAttendees.Add(employeeAttendee);
                     }
 
-                    employeeAttendees.Add(employeeAttendee);
+                    if( employeeAttendees != null )
+                        employeeAttendees.Add(employeeAttendee);
                 }
             }
             #endregion update distributor data
@@ -433,30 +434,34 @@ namespace asi.asicentral.web.Controllers.Show
                                 log.Debug("Index - start updating attendee data");
                                 var showAttendees = ObjectService.GetAll<ShowAttendee>().Where(item => item.ShowId == objShow.Id).ToList();
                                 var attendeesToBeDeleted = showAttendees.Where(attendee => attendee.IsExisting == false).ToList();
-                                for (var i = attendeesToBeDeleted.Count() - 1; i >= 0; i--)
+                                if (attendeesToBeDeleted.Count > 0)
                                 {
-                                    var attendee = attendeesToBeDeleted[i];
-                                    if (attendee.EmployeeAttendees != null && attendee.EmployeeAttendees.Any())
+                                    for (var i = attendeesToBeDeleted.Count() - 1; i >= 0; i--)
                                     {
-                                        for (var j = attendee.EmployeeAttendees.Count() - 1; j >= 0; j--)
+                                        var attendee = attendeesToBeDeleted[i];
+                                        if (attendee.EmployeeAttendees != null && attendee.EmployeeAttendees.Any())
                                         {
-                                            ObjectService.Delete(attendee.EmployeeAttendees[j]);
+                                            for (var j = attendee.EmployeeAttendees.Count() - 1; j >= 0; j--)
+                                            {
+                                                ObjectService.Delete(attendee.EmployeeAttendees[j]);
+                                            }
                                         }
+
+                                        if (attendee.DistShowLogos != null && attendee.DistShowLogos.Any())
+                                        {
+                                            for (var j = attendee.DistShowLogos.Count() - 1; j >= 0; j--)
+                                            {
+                                                ObjectService.Delete(attendee.DistShowLogos[j]);
+                                            }
+                                        }
+
+                                        ObjectService.Delete<ShowAttendee>(attendee);
                                     }
 
-                                    if (attendee.DistShowLogos != null && attendee.DistShowLogos.Any())
-                                    {
-                                        for (var j = attendee.DistShowLogos.Count() - 1; j >= 0; j--)
-                                        {
-                                            ObjectService.Delete(attendee.DistShowLogos[j]);
-                                        }
-                                    }
-
-                                    ObjectService.Delete<ShowAttendee>(attendee);
+                                    showAttendees.ForEach(a => a.IsExisting = false);
+                                    ObjectService.SaveChanges();
+                                    log.Debug(string.Format("{0} company attendees have been deleted for '{0}' after uploading", attendeesToBeDeleted.Count, objShow.Name));
                                 }
-
-                                showAttendees.ForEach(a => a.IsExisting = false);
-                                ObjectService.SaveChanges();
 
                                 // delete any employee attendees not in the sheet
                                 var attendeeIds = ObjectService.GetAll<ShowAttendee>().Where(item => item.ShowId == objShow.Id).Select(a => a.Id).ToList();
@@ -465,7 +470,7 @@ namespace asi.asicentral.web.Controllers.Show
                                 for (var k = attendees.Count - 1; k >= 0; k--)
                                 {
                                     if (employeeAttendees.FirstOrDefault(a => a.EmployeeId == attendees[k].EmployeeId) == null)
-                                    {
+                                    { // delete employee from attendee list only, not from database
                                         countDel++;
                                         ObjectService.Delete(attendees[k]);
                                     }
@@ -474,6 +479,7 @@ namespace asi.asicentral.web.Controllers.Show
                                 if (countDel > 0)
                                 {
                                     ObjectService.SaveChanges();
+                                    log.Debug(string.Format("{0} employee attendees have been deleted for '{0}' after uploading", countDel, objShow.Name));
                                 }
 
                                 log.Debug("Index - end updating attendee data - " + (DateTime.Now - postAddingStart).TotalMilliseconds);
